ch13.

+++ Chapter03/Example03.04/main.go
@@ -3,6 +3,6 @@ package main
 import "fmt"
 
 func main() {
-	var a int8 = 128
+	var a int16 = 128
 	fmt.Println(a)
 }
